Enum quiche::h3::Error

pub enum Error {

    Done,
    BufferTooShort,
    InternalError,
    ExcessiveLoad,
    IdError,
    StreamCreationError,
    ClosedCriticalStream,
    MissingSettings,
    FrameUnexpected,
    FrameError,
    QpackDecompressionFailed,
    TransportError(Error),
    StreamBlocked,
    SettingsError,
    RequestRejected,
    RequestCancelled,
    RequestIncomplete,
    MessageError,
    ConnectError,
    VersionFallback,

}

An HTTP/3 error.

Variants

Done

There is no error or no work to do

BufferTooShort

The provided buffer is too short.

InternalError

Internal error in the HTTP/3 stack.

ExcessiveLoad

Endpoint detected that the peer is exhibiting behavior that causes. excessive load.

IdError

Stream ID or Push ID greater that current maximum was used incorrectly, such as exceeding a limit, reducing a limit, or being reused.

StreamCreationError

The endpoint detected that its peer created a stream that it will not accept.

ClosedCriticalStream

A required critical stream was closed.

MissingSettings

No SETTINGS frame at beginning of control stream.

FrameUnexpected

A frame was received which is not permitted in the current state.

FrameError

Frame violated layout or size rules.

QpackDecompressionFailed

QPACK Header block decompression failure.

TransportError(Error)

Error originated from the transport layer.

StreamBlocked

The underlying QUIC stream (or connection) doesn’t have enough capacity for the operation to complete. The application should retry later on.

SettingsError

Error in the payload of a SETTINGS frame.

RequestRejected

Server rejected request.

RequestCancelled

Request or its response cancelled.

RequestIncomplete

Client’s request stream terminated without containing a full-formed request.

MessageError

An HTTP message was malformed and cannot be processed.

ConnectError

The TCP connection established in response to a CONNECT request was reset or abnormally closed.

VersionFallback

The requested operation cannot be served over HTTP/3. Peer should retry over HTTP/1.1.