quiche::h3

Enum WireErrorCode

pub enum WireErrorCode {

    NoError = 256,
    GeneralProtocolError = 257,
    InternalError = 258,
    StreamCreationError = 259,
    ClosedCriticalStream = 260,
    FrameUnexpected = 261,
    FrameError = 262,
    ExcessiveLoad = 263,
    IdError = 264,
    SettingsError = 265,
    MissingSettings = 266,
    RequestRejected = 267,
    RequestCancelled = 268,
    RequestIncomplete = 269,
    MessageError = 270,
    ConnectError = 271,
    VersionFallback = 272,

}

HTTP/3 error codes sent on the wire.

As defined in RFC9114.

Variants

NoError = 256

No error. This is used when the connection or stream needs to be closed, but there is no error to signal.

GeneralProtocolError = 257

Peer violated protocol requirements in a way that does not match a more specific error code or endpoint declines to use the more specific error code.

InternalError = 258

An internal error has occurred in the HTTP stack.

StreamCreationError = 259

The endpoint detected that its peer created a stream that it will not accept.

ClosedCriticalStream = 260

A stream required by the HTTP/3 connection was closed or reset.

FrameUnexpected = 261

A frame was received that was not permitted in the current state or on the current stream.

FrameError = 262

A frame that fails to satisfy layout requirements or with an invalid size was received.

ExcessiveLoad = 263

The endpoint detected that its peer is exhibiting a behavior that might be generating excessive load.

IdError = 264

A stream ID or push ID was used incorrectly, such as exceeding a limit, reducing a limit, or being reused.

SettingsError = 265

An endpoint detected an error in the payload of a SETTINGS frame.

MissingSettings = 266

No SETTINGS frame was received at the beginning of the control stream.

RequestRejected = 267

A server rejected a request without performing any application processing.

RequestCancelled = 268

The request or its response (including pushed response) is cancelled.

RequestIncomplete = 269

The clientโs stream terminated without containing a fully formed request.

MessageError = 270

An HTTP message was malformed and cannot be processed.

ConnectError = 271

The TCP connection established in response to a CONNECT request was reset or abnormally closed.

VersionFallback = 272

The requested operation cannot be served over HTTP/3. The peer should retry over HTTP/1.1.