quiche::h3

Struct Config

pub struct Config { /* private fields */ }

An HTTP/3 configuration.

Implementations

impl Config

pub const fn new() -> Result<Config>

Creates a new configuration object with default settings.

pub fn set_max_field_section_size(&mut self, v: u64)

Sets the SETTINGS_MAX_FIELD_SECTION_SIZE setting.

By default no limit is enforced. When a request whose headers exceed the limit set by the application is received, the call to the poll() method will return the Error::ExcessiveLoad error, and the connection will be closed.

pub fn set_qpack_max_table_capacity(&mut self, v: u64)

Sets the SETTINGS_QPACK_MAX_TABLE_CAPACITY setting.

The default value is 0.

pub fn set_qpack_blocked_streams(&mut self, v: u64)

Sets the SETTINGS_QPACK_BLOCKED_STREAMS setting.

The default value is 0.

pub fn enable_extended_connect(&mut self, enabled: bool)

Sets or omits the SETTINGS_ENABLE_CONNECT_PROTOCOL setting.

The default value is false.

pub fn set_additional_settings( &mut self, additional_settings: Vec<(u64, u64)>, ) -> Result<()>

Sets additional HTTP/3 settings.

The default value is no additional settings. The additional_settings parameter must not the following settings as they are already handled by this library:

SETTINGS_QPACK_MAX_TABLE_CAPACITY

SETTINGS_MAX_FIELD_SECTION_SIZE

SETTINGS_QPACK_BLOCKED_STREAMS

SETTINGS_ENABLE_CONNECT_PROTOCOL

SETTINGS_H3_DATAGRAM

If such a setting is present in the additional_settings, the method will return the Error::SettingsError error.

If a setting identifier is present twice in additional_settings, the method will return the Error::SettingsError error.