quiche

Struct PathStats

pub struct PathStats {

    pub local_addr: SocketAddr,
    pub peer_addr: SocketAddr,
    pub validation_state: PathState,
    pub active: bool,
    pub recv: usize,
    pub sent: usize,
    pub lost: usize,
    pub retrans: usize,
    pub total_pto_count: usize,
    pub dgram_recv: usize,
    pub dgram_sent: usize,
    pub rtt: Duration,
    pub min_rtt: Option<Duration>,
    pub max_rtt: Option<Duration>,
    pub rttvar: Duration,
    pub cwnd: usize,
    pub sent_bytes: u64,
    pub recv_bytes: u64,
    pub lost_bytes: u64,
    pub stream_retrans_bytes: u64,
    pub pmtu: usize,
    pub delivery_rate: u64,
    pub max_bandwidth: Option<u64>,
    pub startup_exit: Option<StartupExit>,

}

Statistics about the path of a connection.

A connectionโs path statistics can be collected using the path_stats() method.

Fields
local_addr: SocketAddr

The local address of the path.

peer_addr: SocketAddr

The peer address of the path.

validation_state: PathState

The path validation state.

active: bool

Whether the path is marked as active.

recv: usize

The number of QUIC packets received.

sent: usize

The number of QUIC packets sent.

lost: usize

The number of QUIC packets that were lost.

retrans: usize

The number of sent QUIC packets with retransmitted data.

total_pto_count: usize

The number of times PTO (probe timeout) fired.

Loss usually happens in a burst so the number of packets lost will depend on the volume of inflight packets at the time of loss (which can be arbitrary). PTO count measures the number of loss events and provides a normalized loss metric.

dgram_recv: usize

The number of DATAGRAM frames received.

dgram_sent: usize

The number of DATAGRAM frames sent.

rtt: Duration

The estimated round-trip time of the connection.

min_rtt: Option<Duration>

The minimum round-trip time observed.

max_rtt: Option<Duration>

The maximum round-trip time observed.

rttvar: Duration

The estimated round-trip time variation in samples using a mean variation.

cwnd: usize

The size of the connectionโs congestion window in bytes.

sent_bytes: u64

The number of sent bytes.

recv_bytes: u64

The number of received bytes.

lost_bytes: u64

The number of bytes lost.

stream_retrans_bytes: u64

The number of stream bytes retransmitted.

pmtu: usize

The current PMTU for the connection.

delivery_rate: u64

The most recent data delivery rate estimate in bytes/s.

Note that this value could be inaccurate if the application does not respect pacing hints (see SendInfo.at and Pacing for more details).

max_bandwidth: Option<u64>

The maximum bandwidth estimate for the connection in bytes/s.

Note: not all congestion control algorithms provide this metric; it is currently only implemented for bbr2_gcongestion.

startup_exit: Option<StartupExit>

Statistics from when a CCA first exited the startup phase.