Struct quiche::Stats

pub struct Stats {

    pub recv: usize,
    pub sent: usize,
    pub lost: usize,
    pub retrans: usize,
    pub sent_bytes: u64,
    pub recv_bytes: u64,
    pub lost_bytes: u64,
    pub stream_retrans_bytes: u64,
    pub peer_max_idle_timeout: u64,
    pub peer_max_udp_payload_size: u64,
    pub peer_initial_max_data: u64,
    pub peer_initial_max_stream_data_bidi_local: u64,
    pub peer_initial_max_stream_data_bidi_remote: u64,
    pub peer_initial_max_stream_data_uni: u64,
    pub peer_initial_max_streams_bidi: u64,
    pub peer_initial_max_streams_uni: u64,
    pub peer_ack_delay_exponent: u64,
    pub peer_max_ack_delay: u64,
    pub peer_disable_active_migration: bool,
    pub peer_active_conn_id_limit: u64,
    pub peer_max_datagram_frame_size: Option<u64>,
    pub paths: Vec<PathStats>,

}

Statistics about the connection.

A connection’s statistics can be collected using the stats() method.

Fields
recv: usize

The number of QUIC packets received.

sent: usize

The number of QUIC packets sent.

lost: usize

The number of QUIC packets that were lost.

retrans: usize

The number of sent QUIC packets with retransmitted data.

sent_bytes: u64

The number of sent bytes.

recv_bytes: u64

The number of received bytes.

lost_bytes: u64

The number of bytes sent lost.

stream_retrans_bytes: u64

The number of stream bytes retransmitted.

peer_max_idle_timeout: u64

The maximum idle timeout.

peer_max_udp_payload_size: u64

The maximum UDP payload size.

peer_initial_max_data: u64

The initial flow control maximum data for the connection.

peer_initial_max_stream_data_bidi_local: u64

The initial flow control maximum data for local bidirectional streams.

peer_initial_max_stream_data_bidi_remote: u64

The initial flow control maximum data for remote bidirectional streams.

peer_initial_max_stream_data_uni: u64

The initial flow control maximum data for unidirectional streams.

peer_initial_max_streams_bidi: u64

The initial maximum bidirectional streams.

peer_initial_max_streams_uni: u64

The initial maximum unidirectional streams.

peer_ack_delay_exponent: u64

The ACK delay exponent.

peer_max_ack_delay: u64

The max ACK delay.

peer_disable_active_migration: bool

Whether active migration is disabled.

peer_active_conn_id_limit: u64

The active connection ID limit.

peer_max_datagram_frame_size: Option<u64>

DATAGRAM frame extension parameter, if any.

paths: Vec<PathStats>

Statistics of the current paths.