quiche

Struct TransportParams

pub struct TransportParams {

    pub original_destination_connection_id: Option<ConnectionId<'static>>,
    pub max_idle_timeout: u64,
    pub stateless_reset_token: Option<u128>,
    pub max_udp_payload_size: u64,
    pub initial_max_data: u64,
    pub initial_max_stream_data_bidi_local: u64,
    pub initial_max_stream_data_bidi_remote: u64,
    pub initial_max_stream_data_uni: u64,
    pub initial_max_streams_bidi: u64,
    pub initial_max_streams_uni: u64,
    pub ack_delay_exponent: u64,
    pub max_ack_delay: u64,
    pub disable_active_migration: bool,
    pub active_conn_id_limit: u64,
    pub initial_source_connection_id: Option<ConnectionId<'static>>,
    pub retry_source_connection_id: Option<ConnectionId<'static>>,
    pub max_datagram_frame_size: Option<u64>,
    pub unknown_params: Option<UnknownTransportParameters>,

}

QUIC Transport Parameters

Fields
original_destination_connection_id: Option<ConnectionId<'static>>

Value of Destination CID field from first Initial packet sent by client

max_idle_timeout: u64

The maximum idle timeout.

stateless_reset_token: Option<u128>

Token used for verifying stateless resets

max_udp_payload_size: u64

The maximum UDP payload size.

initial_max_data: u64

The initial flow control maximum data for the connection.

initial_max_stream_data_bidi_local: u64

The initial flow control maximum data for local bidirectional streams.

initial_max_stream_data_bidi_remote: u64

The initial flow control maximum data for remote bidirectional streams.

initial_max_stream_data_uni: u64

The initial flow control maximum data for unidirectional streams.

initial_max_streams_bidi: u64

The initial maximum bidirectional streams.

initial_max_streams_uni: u64

The initial maximum unidirectional streams.

ack_delay_exponent: u64

The ACK delay exponent.

max_ack_delay: u64

The max ACK delay.

disable_active_migration: bool

Whether active migration is disabled.

active_conn_id_limit: u64

The active connection ID limit.

initial_source_connection_id: Option<ConnectionId<'static>>

The value that the endpoint included in the Source CID field of a Retry Packet.

retry_source_connection_id: Option<ConnectionId<'static>>

The value that the server included in the Source CID field of a Retry Packet.

max_datagram_frame_size: Option<u64>

DATAGRAM frame extension parameter, if any.

unknown_params: Option<UnknownTransportParameters>

Unknown peer transport parameters and values, if any.

Implementations

impl TransportParams

pub fn to_qlog( &self, owner: TransportOwner, cipher: Option<Algorithm>, ) -> EventData

Available on crate feature qlog only.

Creates a qlog event for connection transport parameters and TLS fields