Crate quiche

Savoury implementation of the QUIC transport protocol.

quiche is an implementation of the QUIC transport protocol as specified by the IETF. It provides a low level API for processing QUIC packets and handling connection state, while leaving I/O (including dealing with sockets) to the application.

Structs

Config
Stores configuration shared between multiple connections.

Connection
A QUIC connection.

Header
A QUIC packet's header.

RangeBuf
Buffer holding data at a specific offset.

Readable
An iterator over the streams that have outstanding data to read.

Stats
Statistics about the connection.

Enums

Error
A QUIC error.

Type
QUIC packet type.

Constants

MAX_CONN_ID_LEN
The maximum length of a connection ID.

VERSION_DRAFT17
The current QUIC wire version.

Functions

accept
Creates a new server-side connection.

connect
Creates a new client-side connection.

negotiate_version
Writes a version negotiation packet.

retry
Writes a retry packet.

Type Definitions

Result