quiche

Function accept

pub fn accept(
    scid: &ConnectionId<'_>,
    odcid: Option<&ConnectionId<'_>>,
    local: SocketAddr,
    peer: SocketAddr,
    config: &mut Config,
) -> Result<Connection>

Creates a new server-side connection.

The scid parameter represents the server’s source connection ID, while the optional odcid parameter represents the original destination ID the client sent before a stateless retry (this is only required when using the retry() function).

Examples:

let conn = quiche::accept(&scid, None, local, peer, &mut config)?;