Function quiche::retry

pub fn retry(
    scid: &[u8], 
    dcid: &[u8], 
    new_scid: &[u8], 
    token: &[u8], 
    out: &mut [u8]
) -> Result<usize>

Writes a stateless retry packet.

The scid and dcid parameters are the source connection ID and the destination connection ID extracted from the received client's Initial packet, while new_scid is the server's new source connection ID and token is the address validation token the client needs to echo back.

The application is responsible for generating the address validation token to be sent to the client, and verifying tokens sent back by the client. The generated token should include the dcid parameter, such that it can be later extracted from the token and passed to the accept() function as its odcid parameter.

Examples:

let (len, src) = socket.recv_from(&mut buf).unwrap();

let hdr = quiche::Header::from_slice(&mut buf[..len], quiche::MAX_CONN_ID_LEN)?;

let token = hdr.token.as_ref().unwrap();

// No token sent by client, create a new one.
if token.is_empty() {
    let new_token = mint_token(&hdr, &src);

    let len = quiche::retry(
        &hdr.scid, &hdr.dcid, &scid, &new_token, &mut out,
    )?;

    socket.send_to(&out[..len], &src).unwrap();
    return Ok(());
}

// Client sent token, validate it.
let odcid = validate_token(&src, token);

if odcid == None {
    // Invalid address validation token.
    return Ok(());
}

let conn = quiche::accept(&scid, odcid, &mut config)?;