Enum quiche::h3::Error

#[repr(C)]

pub enum Error {
    Done,
    BufferTooShort,
    WrongSettingDirection,
    PushRefused,
    InternalError,
    PushAlreadyInCache,
    RequestCancelled,
    IncompleteRequest,
    ConnectError,
    ExcessiveLoad,
    VersionFallback,
    WrongStream,
    LimitExceeded,
    DuplicatePush,
    UnknownStreamType,
    WrongStreamCount,
    ClosedCriticalStream,
    WrongStreamDirection,
    EarlyResponse,
    MissingSettings,
    UnexpectedFrame,
    RequestRejected,
    GeneralProtocolError,
    MalformedFrame,
    QpackDecompressionFailed,
    QpackEncoderStreamError,
    QpackDecoderStreamError,
    TransportError,
}

An HTTP/3 error.

Variants
Done

There is no error or no work to do

BufferTooShort

The provided buffer is too short.

WrongSettingDirection

Setting sent in wrong direction.

PushRefused

The server attempted to push content that the client will not accept.

InternalError

Internal error in the HTTP/3 stack.

PushAlreadyInCache

The server attempted to push something the client already has.

RequestCancelled

The client no longer needs the requested data.

IncompleteRequest

The request stream terminated before completing the request.

ConnectError

Forward connection failure for CONNECT target.

ExcessiveLoad

Endpoint detected that the peer is exhibiting behavior that causes. excessive load.

VersionFallback

Operation cannot be served over HTTP/3. Retry over HTTP/1.1.

WrongStream

Frame received on stream where it is not permitted.

LimitExceeded

Stream ID, Push ID or Placeholder Id greater that current maximum was. used

DuplicatePush

Push ID used in two different stream headers.

UnknownStreamType

Unknown unidirection stream type.

WrongStreamCount

Too many unidirectional streams of a type were created.

ClosedCriticalStream

A required critical stream was closed.

WrongStreamDirection

Unidirectional stream type opened at peer that is prohibited.

EarlyResponse

Inform client that remainder of request is not needed. Used in STOP_SENDING only.

MissingSettings

No SETTINGS frame at beginning of control stream.

UnexpectedFrame

A frame was received which is not permitted in the current state.

RequestRejected

Server rejected request without performing any application processing.

GeneralProtocolError

Peer violated protocol requirements in a way that doesn't match a more specific code.

MalformedFrame

TODO: malformed frame where last on-wire byte is the frame type.

QpackDecompressionFailed

QPACK Header block decompression failure.

QpackEncoderStreamError

QPACK encoder stream error.

QpackDecoderStreamError

QPACK decoder stream error.

TransportError

Error originated from the transport layer.