quiche

Struct Stats

pub struct Stats {

    pub recv: usize,
    pub sent: usize,
    pub lost: usize,
    pub spurious_lost: usize,
    pub retrans: usize,
    pub sent_bytes: u64,
    pub recv_bytes: u64,
    pub acked_bytes: u64,
    pub lost_bytes: u64,
    pub stream_retrans_bytes: u64,
    pub dgram_recv: usize,
    pub dgram_sent: usize,
    pub paths_count: usize,
    pub reset_stream_count_local: u64,
    pub stopped_stream_count_local: u64,
    pub reset_stream_count_remote: u64,
    pub stopped_stream_count_remote: u64,
    pub path_challenge_rx_count: u64,
    pub bytes_in_flight_duration: Duration,

}

Statistics about the connection.

A connectionโs statistics can be collected using the stats() method.

Fields
recv: usize

The number of QUIC packets received.

sent: usize

The number of QUIC packets sent.

lost: usize

The number of QUIC packets that were lost.

spurious_lost: usize

The number of QUIC packets that were marked as lost but later acked.

retrans: usize

The number of sent QUIC packets with retransmitted data.

sent_bytes: u64

The number of sent bytes.

recv_bytes: u64

The number of received bytes.

acked_bytes: u64

The number of bytes sent acked.

lost_bytes: u64

The number of bytes sent lost.

stream_retrans_bytes: u64

The number of stream bytes retransmitted.

dgram_recv: usize

The number of DATAGRAM frames received.

dgram_sent: usize

The number of DATAGRAM frames sent.

paths_count: usize

The number of known paths for the connection.

reset_stream_count_local: u64

The number of streams reset by local.

stopped_stream_count_local: u64

The number of streams stopped by local.

reset_stream_count_remote: u64

The number of streams reset by remote.

stopped_stream_count_remote: u64

The number of streams stopped by remote.

path_challenge_rx_count: u64

The total number of PATH_CHALLENGE frames that were received.

bytes_in_flight_duration: Duration

Total duration during which this side of the connection was actively sending bytes or waiting for those bytes to be acked.